requirements.txt
sounddevice>=0.5.0
numpy>=1.24.0
scipy>=1.10.0
pynput>=1.7.6
pyperclip>=1.9.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
